изменил тип int на BIGINT

+++ data/doctor.py
@@ -8,7 +8,7 @@ class Doctor(SqlAlchemyBase):
 
     id = sqlalchemy.Column(sqlalchemy.Integer, autoincrement=True,
                            primary_key=True)
-    chat_id = sqlalchemy.Column(sqlalchemy.Integer)
+    chat_id = sqlalchemy.Column(sqlalchemy.BIGINT)
     doctor_code = sqlalchemy.Column(sqlalchemy.String(7))
     patient = orm.relation('Patient', back_populates='doctor',
                            passive_deletes='all')
