+++ data/university.py
@@ -6,7 +6,7 @@ class University(SqlAlchemyBase):
     __tablename__ = 'university'
     id = sqlalchemy.Column(sqlalchemy.Integer, autoincrement=True,
                            primary_key=True)
-    chat_id = sqlalchemy.Column(sqlalchemy.Integer)
+    chat_id = sqlalchemy.Column(sqlalchemy.BIGINT)
